requirements.txt
asgiref==3.8.1
Django==5.0.6
mysql==0.0.3
mysql-connector-python==8.4.0
mysqlclient==2.2.4
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1
